Rename and restructure GameForm component
