+++ src/components/GameList.vue
@@ -6,19 +6,12 @@
         :key="game.id"
         v-bind="game" />
     </div>
-    <!-- <div v-if="userCanAddGames">
-      <h4 class="ui horizontal divider">
-        add a game
-      </h4>
-      <GameForm :forTeam="team.slug" />
-    </div> -->
   </div>
 </template>
 
 <script>
 import { mapGetters } from 'vuex';
 import GameCard from '@/components/GameCard';
-import GameForm from '@/forms/GameForm';
 
 export default {
   name: 'GameList',
@@ -29,7 +22,6 @@ export default {
   },
   components: {
     GameCard,
-    GameForm,
   },
   computed: {
     ...mapGetters(['hydratedGame']),
